requirements.txt
torch
torchvision
torchaudio
wandb
matplotlib
pandas
pyyaml
